# Optimize snapshot and report generation threads in `create_snapshots_reports_scorecard.py`