--- a/extras/create_snapshots_reports_scorecard.py
+++ b/extras/create_snapshots_reports_scorecard.py
@@ -15,7 +15,6 @@ Options:
 import distutils.dir_util
 import glob
 import logging
-import math
 import os
 import shutil
 import subprocess