# <!-- Use the title to describe PR changes in the imperative mood --> #

## 🗣 Description ##

This PR gives work to threads more efficiently when creating snapshots and reports in `extras/create_snapshots_reports_scorecard.py`.

Note that this does not include any changes to the way third-party snapshots and reports are generated.  That should be done in a later PR (see issue https://github.com/cisagov/cyhy-reports/issues/60).

I also took this opportunity to improve a few logging statements that did not include the thread name.  This will make the log output easier to understand.

<!-- Describe the "what" of your changes in detail. -->
<!-- To avoid scope creep, limit changes to a single goal. -->

## 💭 Motivation and context ##

The current code splits the list of snapshots/reports to be generating into a group of sub-lists and assigns each thread one of those sub-lists to work on.  This PR improves upon this by enabling each thread to pull one snapshot or report (to be generated) from a single list and generating it, repeating the process until the list is empty, at which point each thread exits.  This is a more efficient way to utilize the threads.  It should result in a faster overall processing time by preventing some threads from having to generate multiple long-running snapshots/reports, while other threads are sitting idle because they have already completed their work.

This PR resolves #62.  Since I was making related changes here, I also included changes to resolve #59.

<!-- Why is this change required? -->
<!-- What problem does this change solve? How did you solve it? -->
<!-- Mention any related issue(s) here using appropriate keywords such -->
<!-- as "closes" or "resolves" to auto-close them on merge. -->

## 🧪 Testing ##

I tested this by commenting out or not executing (via switches) parts of the script that were not relevant (e.g. pausing/resuming the commander, creating the sample report, creating third-party snapshots/reports, etc).  Then I temporarily modified the script so that it ran a simple `sleep` command instead of `cyhy-snapshot` and `cyhy-report`.  This allowed me to run the entire script (e.g. `./create_snapshots_reports_scorecard.py --no-dock --no-log --no-pause cyhy-read-only scan-read-only`) and I confirmed that the threading and list consumption worked as intended.

Sample test run output (sanitized and snipped for clarity; note I also limited the number of reports here to generate to 100 for a quicker test run):
```
2024-01-26 10:29:15,075 INFO - BEGIN
2024-01-26 10:29:15,075 INFO - Building list of reports to generate...
2024-01-26 10:29:15,085 INFO - Building list of snapshots to generate...
2024-01-26 10:33:53,529 DEBUG - 100 snapshots to generate: [u'AAA', u'BBB', u'CCC', ...]
2024-01-26 10:33:53,529 DEBUG - [Thread-1] 100 snapshot(s) left to generate
2024-01-26 10:33:53,529 INFO - [Thread-1] Starting snapshot for: AAA
2024-01-26 10:33:53,562 DEBUG - [Thread-2] 99 snapshot(s) left to generate
2024-01-26 10:33:53,563 INFO - [Thread-2] Starting snapshot for: BBB
2024-01-26 10:33:53,563 DEBUG - [Thread-3] 98 snapshot(s) left to generate
2024-01-26 10:33:53,564 INFO - [Thread-3] Starting snapshot for: CCC
...
2024-01-26 10:33:54,603 INFO - [Thread-1] Successful snapshot: AAA (1.06 s)
2024-01-26 10:33:54,604 DEBUG - [Thread-1] 68 snapshot(s) left to generate
2024-01-26 10:33:54,605 INFO - [Thread-1] Starting snapshot for: YYY
2024-01-26 10:33:54,645 INFO - [Thread-3] Successful snapshot: CCC (1.04 s)
2024-01-26 10:33:54,699 INFO - [Thread-2] Successful snapshot: BBB (1.03 s)
...
2024-01-26 10:33:59,814 INFO - [Thread-1] Successful snapshot: ZZZ (2.04 s)
2024-01-26 10:33:59,814 DEBUG - [Thread-1] 0 snapshot(s) left to generate
2024-01-26 10:33:59,814 INFO - [Thread-1] No snapshots left to generate - thread exiting
...
2024-01-26 10:34:00,023 INFO - Time to complete snapshots: 4.75 minutes
2024-01-26 10:34:00,024 INFO - Longest Snapshots:
2024-01-26 10:34:00,024 INFO - GGG: 2.1 seconds
2024-01-26 10:34:00,025 INFO - HHH: 2.1 seconds
2024-01-26 10:34:00,025 INFO - MMM: 2.1 seconds
...
2024-01-26 10:34:00,026 DEBUG - 100 reports to generate: [u'AAA', u'BBB', u'CCC', ...]
2024-01-26 10:34:00,027 DEBUG - [Thread-33] 100 reports left to generate
2024-01-26 10:34:00,027 INFO - [Thread-33] Starting report for: AAA
2024-01-26 10:34:00,533 DEBUG - [Thread-34] 99 reports left to generate
2024-01-26 10:34:00,534 INFO - [Thread-34] Starting report for: BBB
2024-01-26 10:34:01,037 DEBUG - [Thread-35] 98 reports left to generate
2024-01-26 10:34:01,038 INFO - [Thread-35] Starting report for: CCC
2024-01-26 10:34:01,071 INFO - [Thread-33] Successful report generated: AAA (1.04 s)
2024-01-26 10:34:01,071 DEBUG - [Thread-33] 97 reports left to generate
2024-01-26 10:34:01,071 INFO - [Thread-33] Starting report for: FFF
...
2024-01-26 10:34:02,108 INFO - [Thread-33] Successful report generated: FFF (1.04 s)
2024-01-26 10:34:02,109 DEBUG - [Thread-33] 94 reports left to generate
2024-01-26 10:34:02,109 INFO - [Thread-33] Starting report for: JJJ
2024-01-26 10:34:02,582 INFO - [Thread-34] Successful report generated: BBB (2.05 s)
2024-01-26 10:34:02,583 DEBUG - [Thread-34] 92 reports left to generate
2024-01-26 10:34:02,583 INFO - [Thread-34] Starting report for: NNN
...
2024-01-26 10:34:03,086 INFO - [Thread-35] Successful report generated: CCC (2.05 s)
2024-01-26 10:34:03,087 DEBUG - [Thread-35] 89 reports left to generate
2024-01-26 10:34:03,087 INFO - [Thread-35] Starting report for: PPP
...
2024-01-26 10:34:14,273 INFO - [Thread-47] Successful report generated: ZZZ (2.03 s)
2024-01-26 10:34:14,273 DEBUG - [Thread-47] 0 reports left to generate
2024-01-26 10:34:14,273 INFO - [Thread-47] No reports left to generate - exiting
2024-01-26 10:34:14,274 INFO - Time to complete reports: 0.24 minutes
2024-01-26 10:34:14,274 INFO - Longest Reports:
2024-01-26 10:34:14,274 INFO - QQQ: 2.1 seconds
2024-01-26 10:34:14,274 INFO - TTT: 2.1 seconds
2024-01-26 10:34:14,274 INFO - VVV: 2.1 seconds
...
2024-01-26 10:34:14,275 INFO - Time to complete reports: 0.24 minutes
2024-01-26 10:34:14,275 INFO - Number of snapshots generated: 115
2024-01-26 10:34:14,276 INFO -   Third-party snapshots generated: 0
2024-01-26 10:34:14,276 INFO - Number of snapshots failed: 0
2024-01-26 10:34:14,276 INFO -   Third-party snapshots failed: 0
2024-01-26 10:34:14,276 INFO - Number of reports generated: 100
2024-01-26 10:34:14,276 INFO -   Third-party reports generated: 0
2024-01-26 10:34:14,277 INFO - Number of reports failed: 0
2024-01-26 10:34:14,277 INFO -   Third-party reports failed: 0
2024-01-26 10:34:14,277 INFO - Total time: 4.99 minutes
2024-01-26 10:34:14,277 INFO - END
```

I also ran a test against the entire Production set of orgs and confirmed that everything worked fine (from a thread and work consumption standpoint) with a larger set of snapshots and reports.

<!-- How did you test your changes? How could someone else test this PR? -->
<!-- Include details of your testing environment, and the tests you ran to -->
<!-- see how your change affects other areas of the code, etc. -->

<!--
## 📷 Screenshots (if appropriate) ##

Uncomment this section if a screenshot is needed.

-->

## ✅ Pre-approval checklist ##

<!-- Remove any of the following that do not apply. -->
<!-- Draft PRs should have one or more unchecked boxes. -->
<!-- If you're unsure about any of these, don't hesitate to ask. -->
<!-- We're here to help! -->

- [x] This PR has an informative and human-readable title.
- [x] Changes are limited to a single goal - *eschew scope creep!*
- [x] All relevant type-of-change labels have been added.
- [x] I have read the [CONTRIBUTING](../blob/develop/CONTRIBUTING.md) document.
- [x] These code changes follow [cisagov code standards](https://github.com/cisagov/development-guide).
- [x] All new and existing tests pass.

## ✅ Post-merge checklist ##

<!-- Remove any of the following that do not apply. -->

- [ ] Deploy the updated script to Production.
- [ ] Validate that the updated script runs successfully in Production.