@dav3r 2024-01-26
> It's probably also worth at least creating an issue for more gracefully handling snapshot or report generation failures. Right now I think the corresponding thread will crash and exit its thread loop and we will lose that bit of concurrency. At a minimum we should do some error handling to ensure that the thread does not exit, and eventually we probably want to put the item back onto the queue for another chance before abandoning all hope.

I don't believe we will lose the thread if a failure occurs during snapshot or report generation because those commands (`cyhy-snapshot` and `cyhy-report`) are [run in a sub-process](https://github.com/cisagov/cyhy-reports/blob/c2a99d6b63fbba6c4b4a6072600c02e2c6b85fd2/extras/create_snapshots_reports_scorecard.py#L264-L269), whose [return code is checked by the thread](https://github.com/cisagov/cyhy-reports/blob/c2a99d6b63fbba6c4b4a6072600c02e2c6b85fd2/extras/create_snapshots_reports_scorecard.py#L290-L309).  If I'm misunderstanding you or you think I'm wrong, please let me know.

I think that if a failure occurs, we will end up in the same position as we are now (before this PR).  Namely, we will see in the output that there was one or more failed snapshots/reports and we will have to manually attempt to generate them later.  I will still create an issue to document this where we can discuss options for automatically re-attempting the failed snapshots/reports before giving up.