@dav3r 2024-01-26
> I will still create an issue to document this where we can discuss options for automatically re-attempting the failed snapshots/reports before giving up.

See https://github.com/cisagov/cyhy-reports/issues/107 for this.